+++ SIL.DblBundle/DblMetadata.cs
@@ -1,4 +1,5 @@
 using System;
+using System.IO;
 using System.Xml.Serialization;
 using SIL.DblBundle.Text;
 using SIL.Xml;
@@ -34,17 +35,16 @@ namespace SIL.DblBundle
 		/// </summary>
 		[XmlAttribute("revision")]
 		public string Revision_Surrogate {
-			get { return Revision.ToString(); }
+			get => Revision.ToString();
 			set
 			{
-				int revision;
-				Int32.TryParse(value, out revision);
+				Int32.TryParse(value, out var revision);
 				Revision = revision;
 			}
 		}
 
 		/// <summary>Gets whether bundle is for text</summary>
-		public bool IsTextReleaseBundle { get { return Type == "text"; } }
+		public bool IsTextReleaseBundle => Type == "text";
 	}
 
 	/// <summary>
@@ -55,15 +55,29 @@ namespace SIL.DblBundle
 	public abstract class DblMetadataBase<TL> : DblMetadata, IProjectInfo where TL: DblMetadataLanguage, new()
 	{
 		/// <summary>
-		/// Loads information about a Digital Bible Library bundle from the specified projectFilePath.
+		/// Loads information about a Digital Bible Library bundle from the specified reader.
 		/// </summary>
-		public static T Load<T>(string projectFilePath, out Exception exception) where T : DblMetadataBase<TL>
+		/// <param name="metadataReader">A TextReader object assumed to be positioned at the start of the </param>
+		/// <param name="resourceIdentifier">Filename or other meaningful identifier that can be used in an error message if there is a
+		/// problem reading the data or creating the desired type of DblMetadata object from it.</param>
+		/// <param name="exception">Set if the metadata object could not be instantiated or initialized.</param>
+		public static T Load<T>(TextReader metadataReader, string resourceIdentifier,  out Exception exception) where T : DblMetadataBase<TL>
 		{
-			var metadata = XmlSerializationHelper.DeserializeFromFile<T>(projectFilePath, out exception);
+			string data;
+			try
+			{
+				data = metadataReader.ReadToEnd();
+			}
+			catch (Exception e)
+			{
+				exception = e;
+				return null;
+			}
+			var metadata = XmlSerializationHelper.DeserializeFromString<T>(data, out exception);
 			if (metadata == null)
 			{
 				if (exception == null)
-					exception = new ApplicationException(string.Format("Loading metadata ({0}) was unsuccessful.", projectFilePath));
+					exception = new ApplicationException($"Loading metadata ({resourceIdentifier}) was unsuccessful.");
 				return null;
 			}
 			try
@@ -75,6 +89,15 @@ namespace SIL.DblBundle
 				exception = e;
 			}
 			return metadata;
+		}
+
+		/// <summary>
+		/// Loads information about a Digital Bible Library bundle from the specified projectFilePath.
+		/// </summary>
+		public static T Load<T>(string projectFilePath, out Exception exception) where T : DblMetadataBase<TL>
+		{
+			using (var stream = new FileStream(projectFilePath, FileMode.Open))
+				return Load<T>(new StreamReader(stream), projectFilePath, out exception);
 		}
 
 		/// <summary>